requirements.txt
flask-cors>=5.0.1
flask>=3.1.0
openai>=1.75.0
pandas>=2.2.3
python-dotenv>=1.1.0
openpyxl>=3.1.5
requests
